support multiple OS

+++ System/check-reconfigure.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-if [ -f /boot/firmware/reconfigure ]
+if [ -d /boot/firmware ] && [ -f /boot/firmware/reconfigure ]
 then
     sudo cloud-init clean
     sudo rm /etc/netplan/*
